Move dumpdev_list into mount.py datasource
